Updated to Version 1.2.3
#Additions#

Fixed a few UI errors and cleaned up UI in general.

Added  blue hint text in first few levels to help the user understand that they need to enter values into the white boxes for assigning values to the vectors.

#Attempting to fine tune difficulty#

!!!MOVED LEVEL 6 TO BE LEVEL 2!!!

#Now, the old level 2 is level 3, level 3 is level 4, etc, and level 7 is still level 7.#

Modified Level 4 to have the two stars be slightly farther apart from eachother.

Modified Level 5 to have the three stars close together. Centre of mass is the same as before.

Modified Level 6 to have the two stars slightly closer together.

Adjusted the star in level 2 to generate slightly less gravity.

Adjusted the difficulty of level 7 by making the star move farther up and down, but at a slower speed.

Added a new level (8) where a smaller star is orbiting around a larger star. This level should be much more challenging. (Difficulty on this needs to be tweaked)